refactor: moved classes into separate files
